GC the FOOTERS logic.
The storage of mstate pointers in the footers is an intersting feature,
but implementing them in CHERI would be quite expensive (unbound_ptr()
would need to be able to walk all the per-instance segment lists).

Since the ifdefs are making it hard for me to read the code I'm just
removing them.